Move authed stuff to another middleware
